Bug when using the multiple option
Today I wanted to implement select-2 in combination with Ember. After googling, I found this project. But when I was using the code, something went wrong. Some items seems all of a sudden just to vanish.

So I started to debug the code. Problem is the following:
Using the matchIndex as key does not work (line 148), as at line 150 the value is deleted. By deleting the current value, all the remaining values will go 'one index lower'. 

Next time you enter the loop, the following value has the same index (that is 0) as the previous value. In the end all the items are gone, except the latest. >> Fix is simple, just use the `push-function`.