[WIP] Tagging support with ember data support
I am currently working on supporting tags in this component, while also
supporting ember data. You can pass in an ember data record array
(DS.RecordArray) instance, and it will create new tags based on the
`type` of the recordarray.

This PR is still a WIP, any feedback is welcome!